==> code/04_file_generation.ipynb <==
import pandas as pd
import numpy as np
import os


This notebook is intended for turning standardized raw NLP data into text files for use in the RNNs

### Helper functions

def clean_ingredients(ingredients):
    # necessary due to the way .csv stores data. Lists are converted to a long string. Functions returns and actual list of strings
    # and takes opportunity to remove whitespace and unnecessary characters
    full = []
    ingredients = ingredients.replace('\\n', ',').replace('\n',',')
    for string in ingredients.split(','):
        string = ' '.join(string.strip().split())
        string = string.replace("'", '').replace('\\xa0',' ').replace('[','').replace(']','').replace('"','')
        full += string.split(',')
    return full

def create_text_file(col, path, filename):
    with open (f"{path}/{filename}.txt", 'w', encoding='utf-8') as f:
        for i in col:
            f.write(f"{i}\n")

df = pd.read_csv('../data/generated/tidy_recipe_data_all.csv')

df['ingredients'] = df['ingredients'].apply(clean_ingredients)

df['ingredients'].head()

df.head()

kf = df[df['label'] == 1]
kf.head()

nkf = df[df['label'] == 0]

#kf.reset_index(inplace=True)

nkf.head()

len(kf)

len(nkf)

nkf.index[0]

if not os.path.exists('../data/generated/text_files'):
    os.mkdir('../data/generated/text_files')
    os.mkdir('../data/generated/text_files/train')
    os.mkdir('../data/generated/text_files/train/pos')
    os.mkdir('../data/generated/text_files/train/neg')

for i in list(kf.index):
    create_text_file(kf['ingredients'][i], '../data/generated/text_files/train/pos', i)


    
for i in list(nkf.index):
    create_text_file(nkf['ingredients'][i], '../data/generated/text_files/train/neg/', i)

os.mkdir('test')
